feat: 支持 WSL 观察模式（Windows 观察 WSL 内 Agent 项目）
- platform.rs 新增 WSL 路径映射纯函数（UNC 双向映射、/mnt/c 盘符映射、
  wsl.exe 输出 UTF-16 自适应解码、发行版枚举）
- hook_install.rs 支持 WSL 分支：hook 配置写入 WSL 侧 \\wsl$\... 路径，
  hook 命令用 /mnt/c/.../trellis-card.exe（WSL bash 语法）
- hook_cli.rs 事件入口把 Linux 项目路径归一化为 UNC
- lib.rs archive_task 走 wsl.exe 执行 task.py；新增 get_wsl_info/set_wsl_distro
- 前端设置面板新增「WSL 观察」区块
- 新增 docs/wsl-support.md
- 版本号升至 0.2.2

+++ src-tauri/src/ipc.rs
@@ -212,7 +212,7 @@ where
                     wide.as_ptr(),
                     PIPE_ACCESS_DUPLEX | FILE_FLAG_FIRST_PIPE_INSTANCE,
                     PIPE_TYPE_MESSAGE | PIPE_READMODE_MESSAGE | PIPE_WAIT,
-                    PIPE_UNLIMITED_INSTANCES as u32,
+                    PIPE_UNLIMITED_INSTANCES,
                     BUF as u32,
                     BUF as u32,
                     0,
